refactor: :art: funcs name case fix
