+++ src/helpers/css/getElementHeight.ts
