+++ src/helpers/truncateWord.ts
