+++ src/ui/DragSensible.tsx
@@ -1,4 +1,4 @@
-import { useDragStore } from "@/lib/store/drag-store";
+import { useDragStore } from "@/lib/store/dragStore";
 import { memo, useEffect, useRef, useState } from "react";
 
 // Dump implementation, every element will create a new listener, need to find a way to share a single listener
