+++ src/ui/folders/FolderFooter.tsx
@@ -1,5 +1,5 @@
 import { JazzFolder } from "@/lib/jazz/schema";
-import { useModalStore } from "@/lib/store/modal-store";
+import { useModalStore } from "@/lib/store/modalStore";
 import InlineButton from "../InlineButton";
 import PlusIcon from "@/assets/icons/plus.svg?react";
 import PeopleIcon from "@/assets/icons/people.svg?react";
