+++ src/ui/graph/nodes/graphDrawDialog.ts
@@ -1,7 +1,7 @@
 import { NodeObject } from "react-force-graph-2d";
-import { hexToRgba } from "@/helpers/css/hex-to-rgba";
-import { getCssVariable } from "@/helpers/css/get-css-variable";
-import { truncateWord } from "@/helpers/truncate-word";
+import { hexToRgba } from "@/helpers/css/hexToRgba";
+import { getCssVariable } from "@/helpers/css/getCssVariable";
+import { truncateWord } from "@/helpers/truncateWord";
 import { retrieveLaunchParams } from "@telegram-apps/sdk-react";
 
 export const graphDrawDialog = (
